Move server libs to a separate directory
