--- server/blobService.js
@@ -1,1 +1,0 @@
-module.exports = require("./blobService-s3");
