+++ server/lib/blobService/azure.js
