+++ server/routes/users.js
@@ -1,13 +1,10 @@
 const express = require("express");
-const getUserFromSession = require("../userSessions.js").getUserFromSession;
 const request = require("request");
-const bodyParser = require("body-parser");
 const cryptography = require("crypto");
 
-const broadcast = require("../webSocket.js").broadcast;
-const db = require("../postgresQueries.js").queries;
-const blobService = require("../blobService.js");
-const upload = blobService.upload;
+const db = require("../postgresQueries").queries;
+const getUserFromSession = require("../lib/userSessions").getUserFromSession;
+const broadcast = require("../lib/webSocket").broadcast;
 
 function catchAsyncErrors(fn) {
   return (req, res, next) => {
